Restructure temporal module model
