+++ UD-Viz-Core/src/Modules/Temporal/Model/3DTemporalTileset.js
@@ -1,10 +1,8 @@
-import { TransactionsManager } from '../../Utils/TransactionsManager.js';
+import { TransactionsManager } from '../Utils/TransactionsManager.js';
 import { $3DTemporalVersion } from './3DTemporalVersion.js';
 
 export class $3DTemporalTileset {
     constructor(json) {
-        // TODO: pour l'instant les dates sont des années, les gérer ensuite
-        //  avec moment
         this.startDate = json.startDate;
         this.endDate = json.endDate;
