+++ UD-Viz-Core/src/Modules/Temporal/Model/3DTemporalTransactionAggregate.js
