+++ UD-Viz-Core/src/Modules/Temporal/Model/jsonSchemas/3DTILES_temporal.tileset.schema.json
