+++ UD-Viz-Core/src/Modules/Temporal/Model/jsonSchemas/3DTILES_temporal.versionTransition.schema.json
